OrderService/src/main/java/com/fullstack/orderservice/DomainLogic/OrderLogic.java
package com.fullstack.orderservice.DomainLogic;

import com.fullstack.orderservice.DBAccessEntities.Order;
import com.fullstack.orderservice.Repositories.OrderRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class OrderLogic {

    @Autowired
    private OrderRepository orderRepository;

    public List<Order> getAllOrders(){
        return orderRepository.findAll();
    }

    public Order insertOrder(Order order){
        return orderRepository.save(order);
    }

    public Optional<Order> getOrderByFirstName(String firstName){
        return orderRepository.findByFirstName(firstName);
    }
}